+++ compiler/gen/sql/testdata/golden/bidi_query_pkg.go
@@ -313,7 +313,7 @@ func (q *PostQuery) sqlCount(ctx context.Context) (int, error) {
 		}
 	}
 	spec := q.querySpec()
-	spec.Node.Columns = nil
+	spec.Node.Columns = q.ctx.Fields
 	spec.From = from
 	return sqlgraph.CountNodes(ctx, q.config.Driver, spec)
 }